requirement.txt
Django
line-bot-sdk
psycopg2